Moving the function to the client file is required
