+++ app/(authenticated)/admin/roles/[roleID]/rolesActions.ts
@@ -62,26 +62,3 @@ export async function removePermissionFromRole(
   revalidatePath(`/admin/roles/${roleID}`);
   return { ok: true } as const;
 }
-
-export function exposedUserToUserType(users: ExposedUser[]): UserType[] {
-  let usersReturning: UserType[] = [];
-  users.map((u) => {
-    let tempNickname: string = "",
-      tempAvatar: string = "";
-    if (u.nickname != undefined) {
-      tempNickname = u.nickname;
-    }
-    if (u.avatar != undefined) {
-      tempAvatar = u.avatar;
-    }
-    let tempUser: UserType = {
-      user_id: u.user_id,
-      first_name: u.first_name,
-      last_name: u.last_name,
-      nickname: tempNickname,
-      avatar: tempAvatar,
-    };
-    usersReturning.push(tempUser);
-  });
-  return usersReturning;
-}
